@@ -1,13 +1,11 @@
-'use strict';
-
 /**
  * @name git
  * @description
  * Utility functions for interacting with git
  */
 
-import * as util from 'util';
-import {exec, ExecException} from 'child_process';
+import util from 'node:util';
+import {$, execaCommand} from 'execa';
 
 const GIT_LOG = 'git log';
 const GIT_LOG_GREP_OPTION = "--grep='%s' -E";
@@ -15,7 +13,6 @@ const GIT_LOG_TAG_OPTION = '%s..HEAD';
 const GIT_LOG_LINE_SEPARATOR = '==END==';
 const GIT_LOG_FORMAT = `%H%n%s%n%b%n${GIT_LOG_LINE_SEPARATOR}`;
 const GIT_LOG_FORMAT_OPTION = '--format=%s';
-const GIT_TAG = 'git describe --tags --abbrev=0';
 
 interface GetLogOptions {
   grep?: string;
@@ -27,16 +24,9 @@ interface GetLogOptions {
  * @description
  * Get the lastest tag
  */
-function getLastTag(): Promise<string> {
-  return new Promise<string>((resolve, reject) => {
-    exec(GIT_TAG, {}, (error: ExecException | null, stdout: string) => {
-      if (error) return reject(error);
-
-      resolve(stdout);
-    });
-  }).then((stdout: string): string => {
-    return stdout.replace('\n', '');
-  });
+export async function getLastTag(): Promise<string> {
+  const {stdout} = await $`git describe --tags --abbrev=0`;
+  return stdout.replace('\n', '');
 }
 
 /**
@@ -44,7 +34,9 @@ function getLastTag(): Promise<string> {
  * @description
  * Read all commits watching match pattern since a certain tag
  */
-function getLog(options = {} as GetLogOptions): Promise<Array<string>> {
+export async function getLog(
+  options = {} as GetLogOptions
+): Promise<Array<string>> {
   const {grep, tag} = options;
   const cmdArgs = [util.format(GIT_LOG_FORMAT_OPTION, GIT_LOG_FORMAT)];
 
@@ -58,18 +50,9 @@ function getLog(options = {} as GetLogOptions): Promise<Array<string>> {
 
   const cmd = `${GIT_LOG} ${cmdArgs.join(' ')}`;
 
-  return new Promise<Array<string>>((resolve, reject) => {
-    exec(cmd, {}, (error: ExecException | null, stdout = '') => {
-      let output: Array<string> = [];
-      if (error) {
-        reject(error);
-      } else {
-        output = stdout ? stdout.split(`\n${GIT_LOG_LINE_SEPARATOR}\n`) : [];
-      }
-
-      resolve(output);
-    });
+  const {stdout} = await execaCommand(cmd, {
+    shell: true,
+    stripFinalNewline: false,
   });
+  return stdout ? stdout.split(`\n${GIT_LOG_LINE_SEPARATOR}\n`) : [];
 }
-
-export {getLastTag, getLog};
